:recycle: Split type to use in new class
