+++ elements/lisk-transaction-pool/test/unit/queue_checkers.spec.ts
@@ -1,4 +1,18 @@
-import { Transaction } from '../../src/transaction_pool';
+/*
+ * Copyright © 2020 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ *
+ */
+import { Transaction } from '../../src/types';
 import { wrapTransaction } from '../utils/add_transaction_functions';
 import * as queueCheckers from '../../src/queue_checkers';
 import * as transactionObjects from '../../fixtures/transactions.json';
