+++ elements/lisk-transaction-pool/test/unit/transaction_pool.spec.ts
@@ -14,7 +14,8 @@
  */
 import * as transactionObjects from '../../fixtures/transactions.json';
 import { Job } from '../../src/job';
-import { Transaction, TransactionPool } from '../../src/transaction_pool';
+import { Transaction } from '../../src/types';
+import { TransactionPool } from '../../src/transaction_pool';
 import * as checkTransactions from '../../src/check_transactions';
 import { wrapTransaction } from '../utils/add_transaction_functions';
 import * as queueCheckers from '../../src/queue_checkers';
